+++ lib/elven_gard/network/endpoint/connection.ex
@@ -1,15 +1,33 @@
-defmodule ElvenGard.Network.Endpoint.MessageProcessor do
+defmodule ElvenGard.Network.Endpoint.Connection do
   @moduledoc false
 
   alias ElvenGard.Network.PacketProcessor
   alias ElvenGard.Network.Socket
   alias ElvenGard.Network.SocketHandler
 
-  @type result ::
-          {:cont, Socket.t()}
-          | {:halt, SocketHandler.stop_reason(), Socket.t()}
+  @type result :: {:cont, Socket.t()} | {:halt, SocketHandler.stop_reason(), Socket.t()}
+  @type init_result :: result() | {:cont, Socket.t(), timeout()}
 
   ## Public API
+
+  @spec init(Socket.t(), module()) :: init_result()
+  def init(%Socket{} = socket, socket_handler) do
+    case socket_handler.handle_init(socket) do
+      {:ok, %Socket{} = new_socket} ->
+        {:cont, new_socket}
+
+      {:ok, %Socket{} = new_socket, timeout}
+      when timeout == :infinity or (is_integer(timeout) and timeout >= 0) ->
+        {:cont, new_socket, timeout}
+
+      {:stop, reason, %Socket{} = new_socket} ->
+        {:halt, reason, new_socket}
+
+      _ ->
+        raise "handle_init/1 must return `{:ok, socket}`, `{:ok, socket, timeout}` " <>
+                "or `{:stop, reason, socket}`"
+    end
+  end
 
   @spec process(binary(), Socket.t(), module(), module()) :: result()
   def process(data, %Socket{} = socket, socket_handler, packet_handler) do
@@ -36,6 +54,17 @@ defmodule ElvenGard.Network.Endpoint.MessageProcessor do
 
       value ->
         raise "invalid return value for handle_message/2 (got: #{inspect(value)})"
+    end
+  end
+
+  @spec halt(SocketHandler.stop_reason(), Socket.t(), module()) :: Socket.t()
+  def halt(reason, %Socket{} = socket, socket_handler) do
+    case socket_handler.handle_halt(reason, socket) do
+      {:ok, %Socket{} = new_socket} ->
+        new_socket
+
+      _ ->
+        raise "handle_halt/2 must return `{:ok, socket}`"
     end
   end